+++ tests/show_reboot_cause_test.py
@@ -11,6 +11,7 @@ modules_path = os.path.dirname(test_path)
 sys.path.insert(0, modules_path)
 
 import show.main as show
+import reboot_cause
 
 """
     Note: The following 'show reboot-cause' commands simply call other SONiC
@@ -38,7 +39,7 @@ class TestShowRebootCause(object):
     def test_reboot_cause_user(self):
         expected_output = "User issued 'reboot' command [User: admin, Time: Thu Oct 22 03:11:08 UTC 2020]\n"
 
-        with mock.patch("show.main.readRebootCauseFile", return_value={"comment": "", "gen_time": "2020_10_22_03_14_07", "cause": "reboot", "user": "admin", "time": "Thu Oct 22 03:11:08 UTC 2020"}):
+        with mock.patch("reboot_cause.read_reboot_cause_file", return_value={"comment": "", "gen_time": "2020_10_22_03_14_07", "cause": "reboot", "user": "admin", "time": "Thu Oct 22 03:11:08 UTC 2020"}):
             runner = CliRunner()
             result = runner.invoke(show.cli.commands["reboot-cause"], [])
             assert result.output == expected_output
@@ -50,7 +51,7 @@ class TestShowRebootCause(object):
         REBOOT_CAUSE_WATCHDOG_JSON = """\
 {"comment": "", "gen_time": "2020_10_22_03_15_08", "cause": "Watchdog", "user": "", "time": ""}"""
 
-        with mock.patch("show.main.readRebootCauseFile", return_value={"comment": "", "gen_time": "2020_10_22_03_15_08", "cause": "Watchdog", "user": "", "time": ""}):
+        with mock.patch("reboot_cause.read_reboot_cause_file", return_value={"comment": "", "gen_time": "2020_10_22_03_15_08", "cause": "Watchdog", "user": "", "time": ""}):
             runner = CliRunner()
             result = runner.invoke(show.cli.commands["reboot-cause"], [])
             assert result.output == expected_output
